+++ gluecodium/src/test/resources/smoke/lambdas/output/android/com/example/smoke/StandaloneProducer.java
@@ -2,7 +2,9 @@
  *
  */
 package com.example.smoke;
+import android.support.annotation.NonNull;
 @FunctionalInterface
 public interface StandaloneProducer {
+    @NonNull
     String apply();
 }